# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Mister Moderator API
  description: |-
    API мини-приложения **«Мистер Модератор»** для MAX — помощника старосты студенческой группы:
    напоминания, задания, долги, экзамены, материалы, почта и роли внутри группы.

    ### Авторизация
    Мини-приложение передаёт в каждом запросе заголовок `X-Max-Init-Data` — строку
    `WebApp.initData` из MAX Bridge. Сервер проверяет её HMAC-подпись токеном бота.

    * `STRICT_AUTH=true` — заголовок обязателен и подпись проверяется; иначе ответ `401`.
    * `STRICT_AUTH=false` (демо-режим, по умолчанию) — без заголовка запросы выполняются
      от тестового старосты (`max_user_id = test_user_001`): при `USE_TEST_DATA=true` это
      староста группы `iu7-42b` из `backend/seed/test_data.json`, иначе автоматически создаётся
      «Тест Тестов» в группе `TEST-GROUP-01`. Если заголовок передан, пользователь берётся
      из его поля `user` **без проверки подписи** — так проверяются остальные роли.

    ### Роли
    `starosta` (4) → `zam` (3) → `proforg` (2) → `student` (1). Права каждой роли —
    в ответе `GET /api/roles`. При нехватке прав — `403`.

    ### Формат ошибок
    `{"detail": "<сообщение>"}` или `{"detail": {"error": {"code": "...", "message": "..."}}}`.
    Ошибки валидации тела запроса — `422` в стандартном формате FastAPI.
  version: 1.0.0
paths:
...
  /api/roles/assign:
    post:
      tags:
      - Роли и группа
      summary: Назначить роль участнику
      operationId: assign_role_api_roles_assign_post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignRoleRequest'
            example:
              studentId: '2'
              roleId: zam
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: 'Право `roles.assign`. Нельзя менять роль себе и участникам своего или более высокого
        уровня, нельзя назначить роль не ниже своей. Назначение `starosta` доступно только старосте и
        означает передачу роли: текущий староста становится студентом.'
      security:
      - MaxInitData: []
      - {}
  /api/group/members/{student_id}:
    delete:
      tags:
      - Роли и группа
      summary: Исключить участника
      operationId: remove_group_member_api_group_members__student_id__delete
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: integer
          title: Student Id
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RemovedMemberResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `group.edit` (только староста). Нельзя исключить себя и участника своего или
        более высокого уровня.
      security:
      - MaxInitData: []
      - {}
  /api/group/members/{student_id}/name:
    patch:
      tags:
      - Роли и группа
      summary: Переименовать участника
      operationId: rename_member_api_group_members__student_id__name_patch
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: integer
          title: Student Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RenameMemberRequest'
            example:
              name: Иванов Дмитрий
      responses:
        '200':
          description: Успешный ответ
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RenamedMemberResponse'
        '422':
          description: Ошибка валидации параметров или тела запроса
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
        '401':
          description: Нет заголовка X-Max-Init-Data (при STRICT_AUTH=true) или он невалиден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '403':
          description: Недостаточно прав для операции
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
        '404':
          description: Объект не найден или принадлежит другой группе
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPError'
      description: Право `group.edit`. Задаёт отображаемое имя участника в группе.
      security:
      - MaxInitData: []
      - {}
...
components:
  schemas:
...
    AssignRoleRequest:
      properties:
        studentId:
          type: string
          title: Studentid
        roleId:
          type: string
          enum:
          - starosta
          - zam
          - proforg
          - student
          title: Roleid
      type: object
      required:
      - studentId
      - roleId
      title: AssignRoleRequest
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    OkResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
      type: object
      title: OkResponse
...
    RemovedMemberResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        removed:
          type: string
          title: Removed
      type: object
      required:
      - removed
      title: RemovedMemberResponse
    RenameMemberRequest:
      properties:
        name:
          type: string
          title: Name
      type: object
      required:
      - name
      title: RenameMemberRequest
    RenamedMemberResponse:
      properties:
        ok:
          type: boolean
          title: Ok
          default: true
        oldName:
          type: string
          title: Oldname
        newName:
          type: string
          title: Newname
      type: object
      required:
      - oldName
      - newName
      title: RenamedMemberResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
    HTTPError:
      title: HTTPError
      type: object
      required:
      - detail
      properties:
        detail:
          description: 'Текст ошибки или объект {error: {code, message}}'
          oneOf:
          - type: string
          - $ref: '#/components/schemas/ErrorResponse'
    ErrorResponse:
      title: ErrorResponse
      type: object
      required:
      - error
      properties:
        error:
          type: object
          required:
          - code
          - message
          properties:
            code:
              type: string
              example: permission_denied
            message:
              type: string
              example: Нет права roles.assign
  securitySchemes:
    MaxInitData:
      type: apiKey
      in: header
      name: X-Max-Init-Data
      description: Строка WebApp.initData из MAX Bridge. Обязательна при STRICT_AUTH=true.
servers:
- url: /
  description: 'Текущий хост (через nginx: https://discreetly-enjoyable-koi.cloudpub.ru/)'
- url: http://localhost:8080
...